src/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "MAF Agent with Foundry Service and Direct Model variants"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-projects>=1.0.0b7",
    "azure-ai-inference>=1.0.0b4",
    "azure-identity>=1.21.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]
